+++ backend/app/api/main.py
@@ -1,6 +1,16 @@
 from fastapi import APIRouter
 
-from app.api.routes import items, login, private, users, utils
+from app.api.routes import (
+    dashboard,
+    generate,
+    generations,
+    items,
+    login,
+    private,
+    templates,
+    users,
+    utils,
+)
 from app.core.config import settings
 
 api_router = APIRouter()
@@ -8,6 +18,10 @@ api_router.include_router(login.router)
 api_router.include_router(users.router)
 api_router.include_router(utils.router)
 api_router.include_router(items.router)
+api_router.include_router(templates.router)
+api_router.include_router(generate.router)
+api_router.include_router(generations.router)
+api_router.include_router(dashboard.router)
 
 
 if settings.ENVIRONMENT == "local":