+++ backend/app/core/db.py
@@ -21,7 +21,9 @@ def init_db(session: Session) -> None:
     # This works because the models are already imported and registered from app.models
     # SQLModel.metadata.create_all(engine)
 
-    user = session.exec(select(User).where(User.email == settings.FIRST_SUPERUSER)).first()
+    user = session.exec(
+        select(User).where(User.email == settings.FIRST_SUPERUSER)
+    ).first()
     if not user:
         user_service.ensure_superuser_exists(
             session=session,